Move `completion` into `criocli` package
Both packages have the same purpose so we should merge them into one.
